moving things around and adding missing license headers
