+++ src/services/MCServer/LobbyServer/LobbyServer.ts
@@ -5,8 +5,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import { ConnectionObj } from "../../../ConnectionObj";
-import { IRawPacket } from "../../shared/interfaces/IRawPacket";
+import { ConnectionObj } from "../ConnectionObj";
+import { IRawPacket } from "../IRawPacket";
 import { MSG_DIRECTION, NPSMsg } from "../messageTypes/NPSMsg";
 import { NPSUserInfo } from "../../shared/messageTypes/npsUserInfo";
 import { PersonaServer } from "../PersonaServer/PersonaServer";
