+++ src/services/MCServer/LoginServer/packet.test.ts
@@ -5,13 +5,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import { ConnectionObj } from "../../../ConnectionObj";
+import * as packet from "./packet";
 
-export interface IRawPacket {
-  connectionId: string;
-  connection: ConnectionObj;
-  data: Buffer;
-  localPort: number;
-  remoteAddress?: string;
-  timestamp: Date | number;
-}
+test("retuns a buffer packet from premadeLogin", () => {
+  expect(packet.premadeLogin()).toBeInstanceOf(Buffer);
+});
